requirements.txt
python-telegram-bot
json
logging